rename message filename (#609)
