+++ core/src/sync/message/mod.rs
@@ -2,50 +2,50 @@
 // Conflux is free software and distributed under GNU General Public License.
 // See http://www.gnu.org/licenses/
 
-mod blockhashes;
-mod blockheaders;
-mod blocks;
-mod blocktxn;
 mod capability;
-mod cmpctblocks;
-mod getblockhashesbyepoch;
-mod getblockheaders;
-mod getblocks;
-mod getblocktxn;
-mod getcmpctblocks;
-mod getterminalblockhashes;
+mod get_block_hashes_by_epoch;
+mod get_block_hashes_response;
+mod get_block_headers;
+mod get_block_headers_response;
+mod get_block_txn;
+mod get_block_txn_response;
+mod get_blocks;
+mod get_blocks_response;
+mod get_compact_blocks;
+mod get_compact_blocks_response;
+mod get_terminal_block_hashes;
+mod get_terminal_block_hashes_response;
 mod handleable;
 mod keys;
 mod message;
 mod metrics;
-mod newblock;
-mod newblockhashes;
+mod new_block;
+mod new_block_hashes;
 mod status;
-mod terminalblockhashes;
 mod transactions;
 
 pub use self::{
-    blockhashes::GetBlockHashesResponse,
-    blockheaders::GetBlockHeadersResponse,
-    blocks::{GetBlocksResponse, GetBlocksWithPublicResponse},
-    blocktxn::GetBlockTxnResponse,
     capability::{
         DynamicCapability, DynamicCapabilityChange, DynamicCapabilitySet,
     },
-    cmpctblocks::GetCompactBlocksResponse,
-    getblockhashesbyepoch::GetBlockHashesByEpoch,
-    getblockheaders::GetBlockHeaders,
-    getblocks::GetBlocks,
-    getblocktxn::GetBlockTxn,
-    getcmpctblocks::GetCompactBlocks,
-    getterminalblockhashes::GetTerminalBlockHashes,
+    get_block_hashes_by_epoch::GetBlockHashesByEpoch,
+    get_block_hashes_response::GetBlockHashesResponse,
+    get_block_headers::GetBlockHeaders,
+    get_block_headers_response::GetBlockHeadersResponse,
+    get_block_txn::GetBlockTxn,
+    get_block_txn_response::GetBlockTxnResponse,
+    get_blocks::GetBlocks,
+    get_blocks_response::{GetBlocksResponse, GetBlocksWithPublicResponse},
+    get_compact_blocks::GetCompactBlocks,
+    get_compact_blocks_response::GetCompactBlocksResponse,
+    get_terminal_block_hashes::GetTerminalBlockHashes,
+    get_terminal_block_hashes_response::GetTerminalBlockHashesResponse,
     handleable::{Context, Handleable},
     keys::{Key, KeyContainer},
     message::{handle_rlp_message, msgid},
-    newblock::NewBlock,
-    newblockhashes::NewBlockHashes,
+    new_block::NewBlock,
+    new_block_hashes::NewBlockHashes,
     status::Status,
-    terminalblockhashes::GetTerminalBlockHashesResponse,
     transactions::{
         GetTransactions, GetTransactionsResponse, TransactionDigests,
         Transactions,
